backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
openai>=1.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
langgraph==1.0.1
langchain-core==1.0.1
langchain-openai==1.0.1
langchain>=0.1.0
pandas==2.3.3
langfuse>=2.0.0
